requirements.txt
openai
python-dotenv
streamlit
google-search-results
